+++ cli/cli/command_framework/lowlevel/args/consts_for_test.go
@@ -1,4 +1,4 @@
-package kurtosis_command
+package args
 
 const (
 	arg1Key = "arg1"