+++ cli/cli/commands/config/init/init.go
@@ -1,18 +1,20 @@
 package init
 
 import (
+	"context"
+	"github.com/kurtosis-tech/kurtosis-cli/cli/command_framework/highlevel/set_selection_arg"
+	"github.com/kurtosis-tech/kurtosis-cli/cli/command_framework/lowlevel"
+	"github.com/kurtosis-tech/kurtosis-cli/cli/command_framework/lowlevel/args"
+	"github.com/kurtosis-tech/kurtosis-cli/cli/command_framework/lowlevel/flags"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/command_str_consts"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/helpers/prompt_displayer"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/kurtosis_config"
-	"github.com/kurtosis-tech/kurtosis-cli/commons/positional_arg_parser"
 	"github.com/kurtosis-tech/stacktrace"
 	"github.com/sirupsen/logrus"
-	"github.com/spf13/cobra"
-	"strings"
 )
 
 const (
-	acceptSendingMetricsArg = "accept-sending-metrics"
+	acceptSendingMetricsArgKey = "accept-sending-metrics"
 
 	overrideConfigPromptLabel = "The Kurtosis Config is already created; do you want to override it?"
 
@@ -21,34 +23,42 @@ const (
 	rejectSendingMetricsInput  = "dont-send-metrics"
 )
 
-var allAcceptSendingMetricsValidInputs = []string{acceptSendingMetricsInput, rejectSendingMetricsInput}
-
-var positionalArgs = []string{
-	acceptSendingMetricsArg,
+var validAcceptSendingMetricsArgValues = map[string]bool{
+	acceptSendingMetricsInput: true,
+	rejectSendingMetricsInput: true,
 }
 
-var InitCmd = &cobra.Command{
-	Use:                   command_str_consts.InitCmdStr + " [flags] " + strings.Join(positionalArgs, " "),
-	DisableFlagsInUseLine: true,
-	Short:                 "Initialize the Kurtosis CLI configuration",
-	Long: "Initializes the configuration file that the CLI uses with the given values",
-	RunE:                  run,
+var InitCmd = &lowlevel.LowlevelKurtosisCommand{
+	CommandStr:               command_str_consts.InitCmdStr,
+	ShortDescription:         "Initialize the Kurtosis CLI configuration",
+	LongDescription:          "Initializes the configuration file that the CLI uses with the given values",
+	Args:                     []*args.ArgConfig{
+		set_selection_arg.NewSetSelectionArg(
+			acceptSendingMetricsArgKey,
+			validAcceptSendingMetricsArgValues,
+		),
+	},
+	RunFunc:                  run,
 }
 
-func init() {
-}
-
-func run(cmd *cobra.Command, args []string) error {
-
-	parsedPositionalArgs, err := positional_arg_parser.ParsePositionalArgsAndRejectEmptyStrings(positionalArgs, args)
+func run(ctx context.Context, flags *flags.ParsedFlags, args *args.ParsedArgs) error {
+	didUserAcceptSendingMetricsStr, err := args.GetNonGreedyArg(acceptSendingMetricsArgKey)
 	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred parsing the positional args")
+		return stacktrace.Propagate(err, "Expected a value for non-greedy arg '%v' but none was found; this is a bug in Kurtosis!", acceptSendingMetricsArgKey)
 	}
-	acceptSendingMetricsStr := parsedPositionalArgs[acceptSendingMetricsArg]
 
-	didUserAcceptSendingMetrics, err := validateMetricsConsentInputAndGetBooleanResult(acceptSendingMetricsStr)
-	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred validating metrics consent input")
+	// We get validation for free by virtue of the KurtosisCommand framework
+	var didUserAcceptSendingMetrics bool
+	if didUserAcceptSendingMetricsStr == acceptSendingMetricsInput {
+		didUserAcceptSendingMetrics = true
+	} else if didUserAcceptSendingMetricsStr == rejectSendingMetricsInput {
+		didUserAcceptSendingMetrics = false
+	} else {
+		// If this happens, there's something wrong with the validation being done via KurtosisCommand
+		return stacktrace.NewError(
+			"Encountered an unrecognized 'should send metrics?' input string '%v', which should never happen; this is a bug in Kurtosis!",
+			didUserAcceptSendingMetricsStr,
+		)
 	}
 
 	kurtosisConfigStore := kurtosis_config.GetKurtosisConfigStore()
@@ -74,38 +84,4 @@ func run(cmd *cobra.Command, args []string) error {
 	}
 
 	return nil
-}
-
-// ====================================================================================================
-//                                       Private Helper Functions
-// ====================================================================================================
-func validateMetricsConsentInputAndGetBooleanResult(input string) (bool, error) {
-
-	didUserAcceptSendingMetrics := false
-
-	isValid := contains(allAcceptSendingMetricsValidInputs, input)
-	if !isValid {
-		 return false, stacktrace.NewError(
-			"You have entered an invalid argument '%v'. "+
-				"'%v' to accept sending metrics or "+
-				"'%v' to skip sending metrics",
-			input,
-			acceptSendingMetricsInput,
-			rejectSendingMetricsInput)
-	}
-
-	if input == acceptSendingMetricsInput {
-		didUserAcceptSendingMetrics = true
-	}
-
-	return didUserAcceptSendingMetrics, nil
-}
-
-func contains(s []string, str string) bool {
-	for _, v := range s {
-		if strings.ToLower(v) == strings.ToLower(str) {
-			return true
-		}
-	}
-	return false
 }